Day 06: Fix clippy issue

+++ day06/src/main.rs
@@ -36,11 +36,12 @@ struct Race {
 
 impl Race {
     fn get_number_of_winning_strategies(&self) -> u64 {
-        let a = -1 as f64;
+        let a = -1_f64;
         let b = self.time as f64;
-        let c = -1f64 * self.record as f64;
+        let c = -1_f64 * self.record as f64;
 
-        let x = (((-1f64 * b) + f64::sqrt(b * b - 4f64*a*c)) / (2f64 * a)).floor() as u64 + 1;
+        let x =
+            (((-1_f64 * b) + f64::sqrt(b * b - 4_f64 * a * c)) / (2_f64 * a)).floor() as u64 + 1;
 
         self.time - (x * 2) + 1
     }
